requirements.txt
fastapi
uvicorn
openai
pinecone-client
sqlalchemy
pydantic
python-dotenv

